+++ module.hpp
@@ -1,9 +1,11 @@
 #ifndef module_hpp
 #define module_hpp
 
-bool isPrime(int n);//проверка на простоту
+#include <cstdint>
+
+bool isPrime(int64_t n);//проверка на простоту
 
 //проверяет условия теоремы Ферма для a и p
-bool checkFerma(int a, int p);
+bool checkFerma(int64_t a, int64_t p);
 
 #endif
